+++ src/main/java/podo/odeego/web/auth/config/LoginMember.java
@@ -1,4 +1,4 @@
-package podo.odeego.web.annotation;
+package podo.odeego.web.auth.config;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
