+++ src/main/java/podo/odeego/web/auth/dto/GenerateTokenResponse.java
@@ -1,4 +1,4 @@
-package podo.odeego.web.auth.jwt.dto;
+package podo.odeego.web.auth.dto;
 
 public record GenerateTokenResponse (
 	String accessToken,
